feat: Update default model quantization to float8 and enable PyTorch inference mode for the worker thread.

+++ backend/backend-python/src/pie_backend/__main__.py
@@ -60,7 +60,7 @@ def main(
 
 
     # testing quantization
-    weight_dtype = "int8"
+    weight_dtype = "float8"
 
     config = RuntimeConfig.from_args(
         model=model,
